+++ src/DynamicDict/__init__.py
@@ -1,6 +1,6 @@
 """A dictionary wrapper that allows attribute access and mutation using both dot notation and dictionary-style indexing."""
 
-__version__ = "0.2.0"
+__version__ = "0.2.1"
 
 # For easy import...
 from .DynamicDict import Dynamic