+++ CHANGELOG.md
@@ -3,12 +3,19 @@
 All notable changes to this project will be documented in this file.
 
 ## [Roadmap]
-- Consider renaming the project...
 - Add some(??) interoperability with TypedDict.
 - Add multiply dunders as intersection operation.
 - Add divide dunders as symmetric difference operation.
+- Consider renaming the project...
 
-## [0.2.0]  - 2024-07-27
+## [0.2.1] - 2024-07-27
+### Added
+- Functions that start with a "self" parameter get the parent self prepended when called.
+
+## Fixed
+- Incorrect error message when bad type passed for _strict_typing in constructor.
+
+## [0.2.0] - 2024-07-27
 ### Changed
 - Change class name to `Dynamic` to better match conventions.